log boat combine_forces input and output
Turns out both small inaccuracy when determining the offset between compass and wind sensor installation orientation as well as the wind sensor oscillating a bit leads to problems if moving at high speeds and/or correspondingly low wind speeds. See https://www.desmos.com/calculator/mb34ijyv6m for an example of how 5° of error in calibration or due to oscillation can lead to 45° of error in calculated end result.